[FOR-UPSTREAM] HID: hid-steam: Make sure rumble work is canceled on removal
When a force feedback command is sent from userspace, work is scheduled to pass
this data to the controller without blocking userspace itself. However, in
theory, this work might not be properly cancelled if the controller is removed
at the exact right time. This patch ensures the work is properly cancelled when
the device is removed.

Signed-off-by: Vicki Pfau <vi@endrift.com>