asus-ally-hid: Add joystick LED ring support
Adds basic support for the joystick RGB LED rings as a multicolour LED
device with 4 LEDs. As this uses the software-mode for setting the LED
colours it also sets the MCU-mode for LED's to static + the first RGB
colour on suspend/reboot/shutdown. This is done to prevent user confusion
if the LED's were to not match what was set, and to maintain consistency.

Signed-off-by: Luke Jones <luke@ljones.dev>
(cherry picked from commit ffe452cafea3adc5719a40c7574a4e789322b606)