[FROM-UPSTREAM] drm/amdgpu: read back register after written for VCN v4.0.5
V4: add read-back for non-DPG case. This is for protection
    purpose as it is not used for producton.

On VCN v4.0.5 there is a race condition where the WPTR is not
updated after starting from idle when doorbell is used. Adding
register read-back after written at function end is to ensure
all register writes are done before they can be used.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/12528
Cc: stable@vger.kernel.org

Signed-off-by: David Wu (Ming Qiang) <David.Wu3@amd.com>
Reviewed-by: Mario Limonciello <mario.limonciello@amd.com>
Tested-by: Mario Limonciello <mario.limonciello@amd.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/12528
Link: https://lore.kernel.org/r/20250513221218.654442-1-David.Wu3@amd.com
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>